plat/arm: Introduce corstone700 platform.
This patch adds support for Corstone-700 foundation IP, which integrates both
Cortex-M0+ and Cortex-A(Host) processors in one handy, flexible subsystem.
This is an example implementation of Corstone-700 IP host firmware.

Cortex-M0+ will take care of boot stages 1 and 2(BL1/BL2) as well as bringing
Host out RESET.Host will start execution directly from BL32 and then will jump
to Linux.

It is an initial port and additional features are expected to be added later.

Change-Id: I7b5c0278243d574284b777b2408375d007a7736e
Signed-off-by: Manish Pandey <manish.pandey2@arm.com>