plat/arm: add compile time flag for CORSTONE700 plat
Added a compile time flag PLAT_ARM_CORSTONE700 to change generic values of
Trusted SRAM and timers.These values for all arm platforms are supposed to be
same and this was true for all platforms till now, but corstone700 has different
values.

We can either move these values from arm common to platform specific headers or
keep it in common and live with a compile time flag.

Change-Id: I432874c443a329b86b5f2ee5c0c181e1f8dd75fb
Signed-off-by: Manish Pandey <manish.pandey2@arm.com>